requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
pgvector
SQLAlchemy
psycopg2-binary
phi-assistant
request
pypdf
psycopg[binary]
google-generativeai
crewai
crewai[tools]
crewai_tools
pysbd
